Move ground_conductivity endpoint to ghpghx app, from reo
Also move the default data into ghpghx/models.py
