+++ ghpghx/urls.py
@@ -36,5 +36,6 @@ urlpatterns = [
     re_path(r'^ghpghx/errors/?$', views.errors),
     re_path(r'^ghpghx/help/?$', views.help),
     re_path(r'^ghpghx/(?P<ghp_uuid>[0-9a-f-]+)/results/?$', views.results),
+    re_path(r'^ghpghx/ground_conductivity/?$', views.ground_conductivity),    
 ]+ static(settings.STATIC_URL, document_root=settings.STATIC_ROOT)
