+++ ghpghx/views.py
@@ -41,7 +41,9 @@ from django.http import HttpResponse
 from django.template import  loader
 from django.views.decorators.csrf import csrf_exempt
 from ghpghx.resources import UUIDFilter
-from ghpghx.models import ModelManager
+from ghpghx.models import ModelManager, GHPGHXInputs
+from reo.utilities import get_climate_zone_and_nearest_city
+from reo.src.load_profile import BuiltInProfile
 
 log = logging.getLogger(__name__)
 
@@ -100,3 +102,40 @@ def results(request, ghp_uuid):
         exc_type, exc_value, exc_traceback = sys.exc_info()
         resp = make_error_resp("Error when trying to make_response")
         return JsonResponse(resp, status=500)
+
+def ground_conductivity(request):
+    """
+    GET ground thermal conductivity based on the climate zone from the lat/long input
+    param: latitude: latitude of the site location
+    param: longitude: longitude of the site location
+    return: climate_zone: climate zone of the site location
+    return: thermal_conductivity [Btu/(hr-ft-degF)]: thermal conductivity of the ground in climate zone
+    """
+    try:
+        latitude = float(request.GET['latitude'])  # need float to convert unicode
+        longitude = float(request.GET['longitude'])
+
+        climate_zone, nearest_city, geometric_flag = get_climate_zone_and_nearest_city(latitude, longitude, BuiltInProfile.default_cities)
+        k_by_zone = copy.deepcopy(GHPGHXInputs.ground_k_by_climate_zone)
+        k = k_by_zone[climate_zone]
+
+        response = JsonResponse(
+            {
+                "climate_zone": climate_zone,
+                "thermal_conductivity": k
+            }
+        )
+        return response
+
+    except ValueError as e:
+        return JsonResponse({"Error": str(e.args[0])}, status=400)
+
+    except KeyError as e:
+        return JsonResponse({"Error. Missing": str(e.args[0])}, status=400)
+
+    except Exception:
+        exc_type, exc_value, exc_traceback = sys.exc_info()
+        debug_msg = "exc_type: {}; exc_value: {}; exc_traceback: {}".format(exc_type, exc_value.args[0],
+                                                                            tb.format_tb(exc_traceback))
+        log.debug(debug_msg)
+        return JsonResponse({"Error": "Unexpected error in ground_conductivity endpoint. Check log for more."}, status=500)
