+++ reo/src/ghp.py
@@ -11,24 +11,6 @@ class GHPGHXInputs:
     Map the GHPGHX inputs to the objects in this class
     """
 
-    ground_k_by_climate_zone = {
-                                "1A": 1.029,
-                                "2A": 1.348,
-                                "2B": 0.917,
-                                "3A": 1.243,
-                                "3B": 1.364,
-                                "3C": 1.117,
-                                "4A": 1.023,
-                                "4B": 0.972,
-                                "4C": 1.418,
-                                "5A": 1.726,
-                                "5B": 1.177,
-                                "6A": 0.977,
-                                "6B": 0.981,
-                                "7A": 1.034,
-                                "7B": 1.508
-                                }
-
     def __init__(self, inputs_dict, **kwargs):
         self.heating_thermal_mmbtu_per_hr = inputs_dict["heating_thermal_load_mmbtu_per_hr"]
         self.heating_thermal_kw = list(np.array(self.heating_thermal_mmbtu_per_hr) * MMBTU_TO_KWH)
