+++ reo/views.py
@@ -35,7 +35,7 @@ import uuid
 import copy
 import json
 from django.http import JsonResponse
-from reo.src.load_profile import BuiltInProfile, LoadProfile, get_climate_zone
+from reo.src.load_profile import BuiltInProfile, LoadProfile
 from reo.src.load_profile_boiler_fuel import LoadProfileBoilerFuel
 from reo.src.load_profile_chiller_thermal import LoadProfileChillerThermal
 from reo.models import URDBError
@@ -53,7 +53,6 @@ import pandas as pd
 from reo.utilities import MMBTU_TO_KWH, generate_year_profile_hourly, TONHOUR_TO_KWHT, get_weekday_weekend_total_hours_by_month
 from reo.validators import ValidateNestedInput
 from datetime import datetime, timedelta
-from reo.src.ghp import GHPGHXInputs
 import numpy as np
 
 
@@ -954,40 +953,3 @@ def schedule_stats(request):
                                                                             tb.format_tb(exc_traceback))
         log.debug(debug_msg)
         return JsonResponse({"Error": "Unexpected error in schedule_stats endpoint. Check log for more."}, status=500)
-
-def ground_conductivity(request):
-    """
-    GET ground thermal conductivity based on the climate zone from the lat/long input
-    param: latitude: latitude of the site location
-    param: longitude: longitude of the site location
-    return: climate_zone: climate zone of the site location
-    return: thermal_conductivity [Btu/(hr-ft-degF)]: thermal conductivity of the ground in climate zone
-    """
-    try:
-        latitude = float(request.GET['latitude'])  # need float to convert unicode
-        longitude = float(request.GET['longitude'])
-
-        climate_zone = get_climate_zone(latitude, longitude)
-        k_by_zone = copy.deepcopy(GHPGHXInputs.ground_k_by_climate_zone)
-        k = k_by_zone[climate_zone]
-
-        response = JsonResponse(
-            {
-                "climate_zone": climate_zone,
-                "thermal_conductivity": k
-            }
-        )
-        return response
-
-    except ValueError as e:
-        return JsonResponse({"Error": str(e.args[0])}, status=400)
-
-    except KeyError as e:
-        return JsonResponse({"Error. Missing": str(e.args[0])}, status=400)
-
-    except Exception:
-        exc_type, exc_value, exc_traceback = sys.exc_info()
-        debug_msg = "exc_type: {}; exc_value: {}; exc_traceback: {}".format(exc_type, exc_value.args[0],
-                                                                            tb.format_tb(exc_traceback))
-        log.debug(debug_msg)
-        return JsonResponse({"Error": "Unexpected error in ground_conductivity endpoint. Check log for more."}, status=500)
